data_pipeline/requirements.txt
requests
azure-storage-blob
